feat:add google AI speacher

+++ readyaapp/models.py
@@ -1,13 +1,13 @@
 from django.db import models
 import uuid
-
+import os
 
 class AudioDocument(models.Model):
     STATUS_CHOICES = (
         ("processing", "Processing"),
         ("done", "Done"),
         ("failed", "Failed"),
-        ("pending_payment", "Pending Payment"), 
+        ("pending_payment", "Pending Payment"),
     )
 
     FILE_TYPE_CHOICES = (
@@ -34,6 +34,7 @@ class AudioDocument(models.Model):
 
     mp3_file = models.FileField(upload_to="uploads/mp3/", null=True, blank=True)
 
+
     status = models.CharField(max_length=20, choices=STATUS_CHOICES, default="processing")
 
     # payment_status = models.CharField(
@@ -59,13 +60,20 @@ class AudioDocument(models.Model):
     created_at = models.DateTimeField(auto_now_add=True)
 
     def delete(self, *args, **kwargs):
-        if self.document_file:
-            self.document_file.delete(save=False)
-        if self.upload_image:
-            self.upload_image.delete(save=False)
-        if self.mp3_file:
-            self.mp3_file.delete(save=False)
+        if self.document_file and os.path.isfile(self.document_file.path):
+           self.document_file.delete(save=False)
+
+        if self.upload_image and os.path.isfile(self.upload_image.path):
+           self.upload_image.delete(save=False)
+
+        if self.mp3_file and os.path.isfile(self.mp3_file.path):
+           self.mp3_file.delete(save=False)
+
         super().delete(*args, **kwargs)
+
 
     def __str__(self):
         return f"{self.email} - {self.file_type} - {self.status}"
+
+
+
